Refactor member and fhr loop into a separate function
If a file is not found or loaded properly, we want to stop loading data for just that day, but still continue on to the next day. The best way to do this is to refactor the member and fhr loops into a separate function. When a file cannot be loaded, a LoadingError is raised, and caught within the load_ens_fcst() function, which inserts missing data and skips the current date.