Add Unit Conversion to loading functions
Now when loading obs or forecasts, a unit conversion can be specified so that the resulting data is returned in a given unit. This is necessary for the ensemble spread, which needs to be calculated on the data after the units have been converted.

Also added a log parameter to these functions, which is useful for precipitation, which needs to be log-transformed before the data is returned.