Bugfix - Correctly reshape binary forecast files before yrev flipping
I was reading in the binary forecasts, and if the `yrev` option was on I was calling `np.flipud()` on that data. I never reshaped to 2 dimensions, however, so it was simply reversing the entire 1-d array. I now reshape to 2 dimensions, flip in the y-direction, and then reshape back to 1 dimensions.

This fixes an issue where the data seemed to be shifted by a gridpoint in the x-direction.