Add code to deal with converting a precip rate to mm
Converting a precip rate to mm requires knowledge of the length of each forecast hour period (eg. 6-hourly fhr files, 24-hourly, etc.).

The unit of precip rate is assumed to be kg/m^2/s, and so is transformed to mm by multiplying by the number of seconds in the fhr file period (eg. 86400 for daily files).