Support variables that accumulate over the entire forecast period
For example, ECENS precip accumulates from the very beginning of the forecast, so at each fhr the value of that field is not the accumulation over the past 6 hours, but the total accumulation since the initial time. To calculate the sum between 2 forecast hours, the sum is field_fhr2 minus field_fhr1.